tus/file-hooks/upload-status/requirements.txt
azure-cosmos
psutil
pyzmq